\name{fastgrid-package}
\alias{fastgrid-package}
\alias{fastgrid}
\docType{package}
\title{
\packageTitle{fastgrid}
}
\description{
\packageDescription{fastgrid}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastgrid}
\packageIndices{fastgrid}
% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastgrid}
Maintainer: \packageMaintainer{fastgrid}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
\code{\link[pkg:gstat-package]{gstat}}
}
\examples{
# esimerkki tahan
}
