% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f90kriege.R
\name{fastkriege}
\alias{fastkriege}
\title{Calculate Kriging predictions}
\usage{
fastkriege(
  trend_model = temperature ~ -1,
  data,
  grid,
  cov.pars,
  lsm = NULL,
  lsmy = NULL,
  alt = NULL,
  alty = NULL,
  altlen = 200,
  bg = NULL,
  variable = "temperature",
  LapseRate = 0,
  method = "bilinear"
)
}
\arguments{
\item{trend_model}{formula to build the regression matrices}

\item{data}{data containing station temperatures as SpatialPoints}

\item{grid}{definition of model grid, must contain longitude, latitude coordinates and variables used in trend_model}

\item{cov.pars}{c(sigmasq,phi,tausq)}

\item{lsm, lsmy}{land sea masks for grid and data}

\item{alt, alty}{altitude information for the grid and data}

\item{altlen}{range parameter for altitude (meters)}

\item{bg}{optional background field of the same dimensions as the grid}

\item{variable}{the name of the variable to be gridded, default "temperature"}
}
\description{
Now assumes that the coordinate names are \code{longitude} and \code{latitude}.
}
