% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils.R
\name{grid2points}
\alias{grid2points}
\alias{grid2points_lapserate}
\title{Map grid to poinits}
\usage{
grid2points(grid, data, variable = "temperature")

grid2points_lapserate(grid, data, variable = "temperature",
  LapseRate = MOSget("LapseRate"), modelgrid = NULL, method = "bilinear")
}
\arguments{
\item{grid}{SpatialGridDataFrame}

\item{data}{SpatialPointsDataFrame}

\item{variable}{variable to be mapped, default \code{temperature}}
}
\description{
Map grid to poinits
}
