% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils.R
\name{gridlon}
\alias{gridlon}
\alias{gridlat}
\title{Grid coordinates}
\usage{
gridlon(x)

gridlat(x)
}
\arguments{
\item{x}{spatial grid}
}
\value{
distinct coordinate values
}
\description{
Get longitude or latitude coordinate values of a regular spatial grid.
Uses \code{\link[sp]{getGridTopology}} and  \code{\link[sp]{coordinatevalues}} from the \code{\link[sp]{sp}} package.
}
\examples{

gridlat(MOSgriddata())

}
