% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils.R
\name{Hmat}
\alias{Hmat}
\title{Linear observation operator}
\usage{
Hmat(mlat, mlon, obs, method = "bilinear")
}
\arguments{
\item{mlon, mlat}{model longitude (x) and latitude (y) coordinates}

\item{obs}{matrix with two columns of data x and y coordinates}

\item{method}{interpolation method, default 'bilinear'}
}
\value{
Sparse Matrix on size nrows(obs) times length(mlon)*length(mlat),
each row has at most 4 non-zero elements
}
\description{
Linear observation operator
}
