% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils.R
\name{makelonlatgrid}
\alias{makelonlatgrid}
\title{Generate uniform rectangular longitude-latitude-based SpatialGrid}
\usage{
makelonlatgrid(
  lonmin,
  lonmax,
  latmin,
  latmax,
  londx = 1,
  latdx = 1,
  lon = NULL,
  lat = NULL
)
}
\arguments{
\item{lonmin, lonmax, londx}{longitude min, max and cell size, used if \code{lon} not given.}

\item{latmin, latmax, latdx}{latitude min, max and cell size, used if \code{lat} not given.}

\item{lon, lat}{longitude and latitude grid center points}
}
\description{
Generate uniform rectangular longitude-latitude-based SpatialGrid
}
\examples{
 

sp::getGridTopology(makelonlatgrid(10,20,-10,10))

}
