% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils.R
\name{pointgridding}
\alias{pointgridding}
\title{Grid point data to spatial grid}
\usage{
pointgridding(y = NULL, pointdata, modelgrid, variable = "y", cov.pars)
}
\arguments{
\item{y}{data matrix of point data to be gridded, alternative to giving \code{variable}}

\item{pointdata}{SpatialPointsDataFrame of the spatial points}

\item{modelgrid}{SpatialGrid or SpatialGridDataFrame of the output grid}

\item{variable}{name of the variable in \code{pointdata} to be gridded}

\item{cov.par}{Kriging parameters, sigma^2, phi, tau^2.}
}
\description{
This is an easier to use version of \link{fastkriege}
}
\examples{
 

lonmin <- 19.0; lonmax <- 33.0
latmin <- 59.0; latmax <- 71.5
newgrid <- makelonlatgrid(lonmin=lonmin,lonmax=lonmax,latmin=latmin,latmax=latmax,londx=0.1,latdx=0.1)
nobs <- 50
obs <- data.frame(longitude=runif(nobs,lonmin,lonmax),
                 latitude=runif(nobs,latmin,latmax),
                 temperature=rnorm(nobs,mean=10,sd=4))
sp::coordinates(obs) <- c('longitude','latitude')
out <- pointgridding(pointdata=obs, modelgrid = newgrid, variable="temperature",cov.pars = c(0.5^2,2.0,0.0))
MOSplotting::MOS_plot_field(out,stations=obs,cmin=0,cmax=20)

}
