% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f90kriege.R
\docType{package}
\name{fastgrid}
\alias{fastgrid}
\title{fastgrid: faster Kriging by some f90 code}
\description{
These are some functions here to help with gridding point data to a regular grid.
}
\section{functions}{

See \code{\link{fastkriege}}.

#' @seealso
\code{\link[gstat]{gstat}} \code{\link[MOSfieldutils]{MOSfieldutils}}
}

